chromeos-config: Allow a SKU ID of -1
The absense of a SKU ID is actually a useful thing to match on in at least
one case. Fizz has models with SKU IDs but the original device did not
have one. We still want to match it.

Update the validator to permit a SKU ID of -1. This can be done in the
config by putting the value in brackets:

	simple-sku-map = < (-1) &fizz >;

BUG=chromium:785366
TEST=emerge-fizz -q --nodeps chromeos-config-bsp
See that it allows this

Change-Id: Ie47c8202c8bec0faf570af58f547503dbc290b3f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/772391
Tested-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>