vm_tools: launcher: use control socket to stop crosvm
crosvm provides a control socket that can be used to stop
a running VM more cleanly than SIGTERM/SIGKILL. Use this
as a nicer fallback in the event that maitred is unable
to shutdown in a timely manner.

Since maitred now attempts a cleaner guest shutdown, also
extend the timeout for the shutdown command to give it a
better chance to complete successfully.

BUG=chromium:770404
TEST=`vm_launcher stop all` doesn't leave stray crosvm instances

Change-Id: Ib476e090b451c278910261054ebd829013953f4f
Reviewed-on: https://chromium-review.googlesource.com/773664
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>