chromeos-config: Change use of ${} to {}
I have been thinking about this for a long time. The ${} syntax matches
shell and is natural for things like ${FILESDIR]. But there is really no
need for the dollar sign and the filename templates looks better without
it.

In fact at present the '$' is already optional since we use Python's
string formatting to process the templates.

Drop the dollar sign from the tests and examples. Once we drop it from
model config we can come back and remove it from the schena.

Change-Id: Ieeadc7433104d7fce3215662994082fd7ff18549
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/777443
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>