chromeos-config: Allow disabling of firmware for a model
Sometimes a model does not have valid firmware, or it has firmware but the
signer has no key.

Add a new 'no-firmware' property which tells pack_firmware not to include
a model in the firmware update for a build.

This CL is entirely within chromeos-config since the firmware packer gets
all its instructions from libcros_config_host now.

BUG=chromium:786519
TEST=FEATURES=test sudo -E emerge --nodeps  chromeos-config-host

Change-Id: Iba01c926d27613aaa04765bfed43b3cd4bac5cab
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/777687
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>