power: Add Moshi USB-C to 3.5mm Adapter/Charger to autosuspend whitelist
Moshi's USB-C to 3.5mm Adapter/Charger is available:
https://store.google.com/us/product/moshi_usb_c_headphone_adapter_charger

The dongle may be attached to Chromebooks.
We want to exercise the low power mode of this device, so let's
whitelist it for USB autosuspend.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=b:64950327
TEST=Attach Moshi USB-C to 3.5mm Adapter/Charger to Electro/Reef. Check in
/sys/bus/usb/devices that the device in question's power/runtime_status
settles on "Suspended" and power consumption is ~2.5mA as measured by
an inline meter.

Change-Id: I920dc4c6a86ce65c2a9bf8dfa7ba38862d53a669
Reviewed-on: https://chromium-review.googlesource.com/780724
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>