permission_broker: Add firewall functionality.
With ambient capabilities we can launch 'iptables' from permission_broker,
so we don't really need firewalld anymore. Having two processes where one
suffices is silly.

This code is:

1-A subset of the IpTables code in firewalld, and
2-Not yet used in permission_broker.

There was a lot of duplication between permission_broker::PortTracker and
firewalld:IpTables so this version of the code removes all of this duplication.
The new Firewall class knows how to call the 'iptables' and 'ip' binaries, and
that's about it.

I have a draft change that modifies PortTracker to use the Firewall class. The
current Firewall class interface should be enough to implement all of firewalld
in permission_broker.

BUG=chromium:787578
TEST=New unit tests pass.

Change-Id: I16b82befb9a1eba703142689f52c5b6490928ea9
Reviewed-on: https://chromium-review.googlesource.com/783264
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>