hammerd: Remove fast version check on boot.
Originally, in order to speed up the no-op case on boot, we only check
the RW version at the beginning. (implemented in CL:651788) But we
forgot to check the touchpad firmware version. So the firmware won't
be updated when there is a new version or the firmware is broken.

Because CL:780941 changed the implementation of usb_utils.h, now the
usb device initialization and manipulation is fast. This CL removes
fast check on boot. Now the touchpad firmware can be updated on boot.
We only skip the pairing process on boot.

Now the timing on boot if no update:
- Trigger upstart job: 91 ms
- Run hammerd: 16 ms
- Total: 107 ms

BUG=b:69347600
TEST=FEATURES=test emerge-poppy hammerd
TEST=Manually run hammerd at device

Change-Id: Id81cc385496565966dbf13d0a58ddfb2f21d31e1
Reviewed-on: https://chromium-review.googlesource.com/784755
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>