power: Add AiAiAi C60 Cable to autosuspend whitelist
AiAiAi C60 Cable is available:
https://aiaiai.dk/headphones/tma-2/parts/cables/c60

The cable may be attached to Chromebooks.
We want to exercise the low power mode of this device, so let's
whitelist it for USB autosuspend.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=b:64950327
TEST=Attach AiAiAi C60 Calbe to Electro/Reef. Check in
/sys/bus/usb/devices that the device in question's power/runtime_status
settles on "Suspended" and power consumption is ~2.5mA as measured by
an inline meter.

Change-Id: I3599e663a1b88a511ef7127880e3582457bb0a2b
Reviewed-on: https://chromium-review.googlesource.com/786347
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>