authpolicy: Handle machine_domain and machine_ou parameters
The 'machine_domain' parameter specifies the domain where the machine is
joined to. If empty, joining user's domain is used. This allows to
support situations where the joining admin is in a different domain than
the target domain.

The 'machine_ou' specifies an organizational unit (OU) where the machine
is put into. Without this, an admin has to move a machine manually from
the default 'Computers' OU to the target OU.

Refactors the SambaInterface code to handle situations where user and
machine domain (aka realm) are not the same. In particular, introduces
the 'AccountData' concept that keeps track of server names, IP addresses
etc. for the machine and user account. This concept might be refactored
into a separate class in the future.

BUG=chromium:765411
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy

Change-Id: Ib158932dad343de1d683e664f389f19ac126bb7e
Reviewed-on: https://chromium-review.googlesource.com/789331
Commit-Ready: Roman Sorokin <rsorokin@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Tested-by: Roman Sorokin <rsorokin@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>