installer: Remove an old FS corruption patch for version 10 and lower
We do not needs support for this patching anymore. Besides, this patching needs
the source image version and we are in the process of depricating passing
lsb-release of the source image to the postinst. So it is a good place to prune
this code.

BUG=chromium:786225
TEST=unittest pass; precq pass;

Change-Id: I82726850d8d5eab3e4933723c009a7396fb64669
Reviewed-on: https://chromium-review.googlesource.com/791455
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>