chromeos-config: Don't assume a default node has a given path
We cannot assume that a default node has any particular path that we look
for. At present the code assumes that if there is a default node, then
every path in that node will have values. This is not necessarily true.

Update the code to handle this case.

BUG=b:69386876
TEST=manually on coral
sudo emerge chromeos-config-host
emerge-coral chromeos-bsp-coral-private

Change-Id: Iec4f68b64a9b15bd21a648012496303f5eff4d8b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/794174
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>