u2fd: fix CONT interleave bug
If a CONT frame is not for the currently processed transaction (ie it
has the wrong channel ID), we need to reject it immediately.

BUG=b:70018223
TEST=emerge-eve u2fd u2f-ref-code
on Eve, run the HIDTest tool from the updated u2f-ref-code package
and see it fail on the new 'test_Interleave' test case, then see it succeed
with this additional fix.

Change-Id: If3aed8cc34684a04e24c640de548858fcbd44916
Reviewed-on: https://chromium-review.googlesource.com/803281
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>