libpasswordprovider: Convert libpasswordprovider to shared lib from static lib
Also fix a bug to account for the MADV_UMERGEABLE madvise option not
being available on some kernels.

Note that the main reviews for these changes were done on
https://chromium-review.googlesource.com/#/c/chromiumos/platform2/+/798399/

The changes to the shared lib only were pulled into a separate CL in
order to land it without breaking the build.

BUG=chromium:386606
TEST=run unit tests, manual testing on kevin using test image

Change-Id: I87deae9c5fbf3df2145252b8edc34f1285f43888
Reviewed-on: https://chromium-review.googlesource.com/806854
Commit-Ready: May Lippert <maybelle@chromium.org>
Tested-by: May Lippert <maybelle@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>