hal_adapter: allow board specific setup hooks
Allow the camera-halv3-adapter upstart script to run a per-board
setup-hooks.sh shell script.  This is used to set up, for example,
the EEPROM file that stores the tuning data of the ov13858 sensor
on Soraka, so that the camera HAL can access the file and read the
tuning data.

BUG=b:70183874
TEST=Deploy on Soraka and make sure the EEPROM file's group and
     permission is modified with a Soraka-specific setup-hooks.sh.

Change-Id: I5cbaf57fab36fc571b50f3f4ce0bb0f61a427818
Reviewed-on: https://chromium-review.googlesource.com/807669
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>