cryptohome: Add journaled look up table for PIN sign in
This patch introduces a simple key-value lookup table for use in
low-entropy credential (e.g PIN) sign in. The values are stored
in versioned files in a directory. The class loads pre-existing tables
into memory on initialization, if the relevant directory exists.

The keys are represented as uint64_t. The values are represented as
binary data.

More details can be found in the class comments.

BUG=chromium:794010
TEST=cryptohome unit tests pass

Change-Id: Ie259173c2f9b73f2c09f3c2b57e369d54803c868
Reviewed-on: https://chromium-review.googlesource.com/821351
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>