power: use separate rules file for Intel HDA
With CL:565448 audio power save mode is enabled statically at boot time,
independently of the power source. This exposes an issue on some Chromeboxes
with unstable output on the ALC283 audio when idle, which results in a popping
or humming noise on external speakers.

The split allows us to skip the installation of the rule that enables audio
power save mode on the affected platforms, while the problem on the audio side
is being investigated.

BUG=chromium:775486, b:69234606
TEST=see CQ-DEPEND CL
CQ-DEPEND=CL:826568

Change-Id: Ia22de4853b759dbb55983335c8e5b726bf66627f
Reviewed-on: https://chromium-review.googlesource.com/827167
Commit-Ready: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>