secure_erase_file: Do not write \0 to /proc/sys/vm/drop_caches
The previous code writes '3', '\n', and '\0' to drop_caches because
sizeof("3\n") is defined to be 3. Since '\0' is nether a digit nor a
whitespace character, it is invalid for this sysctl entry. Actually,
on the 3.14 kernel this WriteFile() only writes the first 2 bytes.
This CL changes the code to write just a single character '3'.

Note that there is no practical or observable problem in the old code.
The first two bytes were always written and triggered the cache drop
as intented.

BUG=chromium:794826
TEST=`strace secure_erase_file` not getting EINVAL on veyron_minnie (3.14)
TEST=`secure_erase_file` still triggers "kernel drop_caches:3" kernel log

Change-Id: Ie78624ee8bb7c97ee3019a66af56958bfd691db9
Reviewed-on: https://chromium-review.googlesource.com/826205
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>