common-mk: export SRC so tests can access data files directly
Currently we force tests to copy any inputs they need for tests to the
output directly.  This is a bit wasteful, especially if the inputs are
a bit on the larger side, so set up the canonical SRC variable (that is
pointing to the source tree) for tests to access.

BUG=chromium:704720
TEST=precq passes, as does new imageloader tests
CQ-DEPEND=CL:830215

Change-Id: Id196f4b0d51e11014907aae1fa7677e564038839
Reviewed-on: https://chromium-review.googlesource.com/830195
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>