cryptohome: perform user sign-in effects on mount
When a homedir is mounted/unmounted, perform user sign-in/sign-out
side effects. As of now the effects consists of calling SetUserType,
which controls setting CCD password for Cr50 and is a NOP for other
TPMs.

The sign-in effects are put in a separate Mount method. Note that

1) Though Mount is not the right place for sign-in effects, they are
already performed there. chromium:795310 is created to eventually
address this issue and reach proper separation of concerns.

2) Using anything but a Mount method would make things worse (similar
to how an alternative approach in CL:827613) is worse than this one.
E.g. passing a hook that'd call a method from upper layers would require
passing this hook to MountFactory, which is used to create Mounts.
MountFactory is created in Service and HomeDirs, HomeDirs are in turn
created not only in Service but in Mount itself. Thus a simplistic
approach with Mount::UserSignInEffects() method was chosen as an
adequate solution until Mount/Service refactoring.

See BUG for more details on the workflow.

BUG=b:69972868
TEST=unit tests, checking logs after sign-ins/sign-outs

Change-Id: I372711c11c28bba69565ecf3ab203a63d9683dce
Reviewed-on: https://chromium-review.googlesource.com/831063
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>