chromeos-config: Add support for firmware files in tarballs
This updates the get-touch-firmware-files to support obtaining touch
firmware files from BCS. It still supports them being in ${FILESDIR}.

This works as follows:

- SRC_URI is populated with a call to cros_config_host get-bsp-uris
- Source filenames are obtained by cros_config_host get-touch-firmware-files

The output of the latter is updated to include three lines per file:

- source filename
- destination filename (in /opt/google/touch/firmware)
- symlink filename (in /lib/firmware)

The source filename is either in the tarfile downloaded from BCS (if a bcs
property is provided by the config) or in the ebuild's files/ directory.

Furture work should look at using a similar mechanism for the firmware
packer.

BUG=chromium:779438
TEST=sudo emerge -q --nodeps chromeos-config-host
emerge-reef chromeos-bsp-reef-private
See that it picks up some files from tar file and some from filesdir

Change-Id: I953467cc268a83560e72f7ba36e0bbb2ec2aca36
Signed-off-by: Simon Glass <sjg@chromium.org>