shill: Fix timeout issue is dns_client
The |ares_options.timeout| field specifies the amount of time that one
nameserver is given to respond to a query. |DnsClient.timeout_ms_|
specifies the total time for ALL retries on a query (inclusive of all
DNS servers). |DnsClient.timeout_ms_| is used to set a separate
timeout with the event loop dispatcher for the whole transaction,
which aborts the lookup.

Currently the |DnsClient.timeout_ms_| option is passed into c-ares
as-is, for one lookup using one DNS server.  This means that if the
first DNS lookup attempt times out in c-ares, the event loop timeout
fires shortly thereafter, preventing DnsClient from trying other DNS
servers on the list.

This CL lowers |ares_options.timeout| so that if there are N DNS
servers configured, it uses a timeout of (DnsClient.timeout_ms_ / N).
Separately, it also hardcodes 8 seconds as the default DNS timeout for
connection diagnostics and HTTP requests.

BUG=b:63620574
TEST=cros_workon_make --board=gale --test shill
Verified by blocking first two servers (8.8.8.8 and 8.8.4.4) in the
DHCP server.

Change-Id: I2044026d3d00bcf14cae6a4cfc0d03d933c9df1a
Reviewed-on: https://chromium-review.googlesource.com/836768
Commit-Ready: Reddeiah Raju Konduru <raju.konduru@globaledgesoft.com>
Tested-by: Reddeiah Raju Konduru <raju.konduru@globaledgesoft.com>
Reviewed-by: Kishan Kunduru <kkunduru@chromium.org>