shill: cellular: add optional authentication field to APN database
Some APNs require authentication where a username and password are
provided through an authentication method such as PAP or CHAP. When an
authentication method is not specified, ModemManager currently defaults
to use PAP as the authentication method. Though either PAP or CHAP is
typically accepted as an authentication method, sometimes CHAP is
allowed but PAP is disallowed. This CL adds an optional authentication
field to the MobileAPN message defined in the mobile operator database.

BUG=b:68680371
CQ-DEPEND=CL:857585
TEST=Run unit tests.

Change-Id: Ia52d52b848367013b35027abb2d0cf7bbca1683f
Reviewed-on: https://chromium-review.googlesource.com/857742
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>