chromeos-config: Split the CrosConfig impl into its own class
We want a call to CrosConfig() to be able to return either a DTB or YAML
implementation. We want to have exactly the same methods in each impl, but
the construction of the object is different.

The easiest way to do this seems to be to make CrosConfig a function,
which returns the implementation, CrosConfigImpl. For the most part this
is transparent to users.

We could adjust GetTargetDirectories() to be a top-level function in
libcros_config_host but then it would no-longer be a member of the object
returned by CrosConfig(), which most users expect.

Rename the class, add CrosConfig() as a new function and update users.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge  --nodeps chromeos-config-host

Change-Id: I627895c6183a1a95b2126f8381aa33971c26a745
Signed-off-by: Simon Glass <sjg@chromium.org>