chromeos-config: Split out the FDT impl into its own class
In preparation for making the YAML impl a subclass of CrosConfigImpl,
split out the FDT impl into its own subclass.

There are still references to phandles in the base class. These are not
used with YAML. Future work will move these into the subclass (once we
have the YAML impl working).

For now we stil use the existing YAML code.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge  --nodeps chromeos-config-host

Change-Id: Ic3434b82b1bcf2f036e2c02b1091e17507d957f7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/842469
Reviewed-by: Jason Clinton <jclinton@chromium.org>