shill: manager: add age for geolocation objects
Adds some utility functions to serialize and deserialize TimeTicks
to the GeolocationInfo map so that we can report the age of the
geolocation objects. This allows us to pull the age in from
wpa_supplicant when we create wifi endpoints and report that as
part of the geolocation object, so that we can see if the geo
information is stale.

BUG=chromium:217554
TEST=unit tests, deploy and call D-Bus methods

Change-Id: I1a80c234fef52535664dded6d4fd8b5532630b2b
Reviewed-on: https://chromium-review.googlesource.com/846423
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>