cryptohome: Add class for sign-in hash tree
This patch introduces a class to handle the storage
and management of the low entropy credential metadata.

A hash tree is used as the data structure. This class
utilizes a PersistentLookupTable as a back-end to
store the credential metadata (which occupy the leaf nodes
of the hash tree) on disk.

This class also uses a file to store the hashes of the inner
nodes of the hash tree. This file is called the HashCache file.

There is support for the following operations in the SignInHashTree:
- GetLabel
- StoreLabel
- RemoveLabel

This patch also adds a PersistentLookupTableMock class
which mocks the functionality of the underlying PersistentLookupTable
for unit tests.

BUG=chromium:794010
TEST=cryptohome unit tests pass

Change-Id: I5c43a3d83ac5733ab32114775d3cbf25436d9d33
Reviewed-on: https://chromium-review.googlesource.com/846661
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>