chromeos-config: Adjust identity code to be more C++-ish
The current code accesses the string table directly to pull out the two
strings that it needs. The table is typically fairly small (10 entries) so
the cost of extracting it into a C++ vector is also fairly small:

- we now allocate all the strings, whereas previously we just did one big
    alloc
- we scan the table twice, calling strlen() on each string, as before
- now when accessing a string, we don't need to scan the table each time

Overall it suggests that this CL will not slow things down appreciably.
Empirical tests (with 100k runs) on coral show a time of 4840ms for the
new code and 5010ms for the old code.

Adjust the code to use a vector, for better readability. Rename
GetString() at the same time, for clarity.

BUG=chromium:796267, chromium:798771
TEST=FEATURES=test sudo -E emerge --nodeps chromeos-config-tools

Change-Id: I27e6777fcec54b71887e268f28bd38845b70f202
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/848642
Reviewed-by: Ben Chan <benchan@chromium.org>