rockchip: Don't wait for params buf when it not ready
The params buf might not be ready to dequeue when output buf done.
Waiting for it would affect the performance, and it's unnecessary since
the params would actually take effect on the next frame.

Open v4l2 dev in nonblock mode, and stop trying when dqbuf returns
EAGAIN to avoid the waiting.

BUG=b:69276988
TEST=Build on scarlet

Change-Id: I508e4e0f419f5df1a768981924acda2a73c0a02d
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/848616
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>