camera3_test: support non-standard color bars pattern
Not every sensor supports the color bars test pattern defined by the
Android standard. With this change we consider the test passed if the
captured frame matches any listed pattern. Partners as well as
developers can add a pattern when it is required in the future projects.

BUG=b:65333318
TEST='emerge-soraka arc-camera3-test'
     'arc_camera3_test
        --gtest_filter=Camera3FrameTest/Camera3FrameContentTest.CorruptionDetection/*'

Change-Id: I432ddd6028f76b78d09ed0826ebc3d828904a6ad
Reviewed-on: https://chromium-review.googlesource.com/848876
Commit-Ready: Hung-yu Wu <hywu@chromium.org>
Tested-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>