chromeos-config: Suppress logging by default
libcros_config prints lots of debugging log messages as it opens the
model config, parses it, looks up keys, etc.  This is very useful for
people working on the config or cros_config, but messes up programs and
tests that need to parse the output of library clients.

To fix this, introduce CROS_CONFIG_LOG(), which works just like LOG()
except that it only emits messages when the CROS_CONFIG_DEBUG
environment variable is set to a non-empty value.  The purpose of using
an environment variable instead of a command-line switch or a change to
the interface is so that any client of the library can be debugged
without needing specific awareness of configuring libcros_config
logging.

Change all LOG() calls to CROS_CONFIG_LOG().  VLOG() calls are left
alone because they already require command-line parameters to be
emitted.

BUG=b:70805256
TEST=cros_config and power_supply_info with/without CROS_CONFIG_DEBUG.

Change-Id: I79250616af970dde8531b70c718dece8431776f1
Reviewed-on: https://chromium-review.googlesource.com/850946
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Patrick Berny <pberny@chromium.org>