chromeos-config: Avoid printing premature errors
At present we try to obtain the config in several ways, printing warnings
or errors along the way. Thus, even if eventually we succeed, we can print
error messages. This is confusing.

Adjust the code to collect the errors it finds and display them them at
the end, when we definitely know that no config could be found. Otherwise
the messages are not displayed.

BUG=chromium:785609
TEST=manual test on robo, adjusted to appear as a nasher360:
Before:
CROS_CONFIG_DEBUG=1 cros_config / brand-code
[1202/092111:ERROR:cros_config.cc(87)] The path /chromeos/models/nasher360
	/audio/main does not exist.
bxtda7219max

With this CL, that message is gone

Change-Id: Id2305d8e14b1db2ff0f1a480c81340d1b04f1c51
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/852574
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Patrick Berny <pberny@chromium.org>