shill: disable auto-connect for cellular before a user logs in
shill may manage multiple user profiles and a service may be configured
in one of the user profiles, or in the default profile, or in a few of
them. However, the AutoConnect property of the same service is not
synchronized across multiple profiles, and thus may have a different
value depending on which profile is used at a given moment. If one user
enables auto-connect on a service while another user disables
auto-connect on the same service, it becomes less clear whether
auto-connect should be enabled or not before any user has logged in.
This is particularly problematic for cellular services, which may incur
data cost. To err on the side of caution, this CL modifies shill to
temporarily disable auto-connect for cellular before a user session has
started.

BUG=b:70835725
TEST=Run unit tests.
TEST=Test the following on a DUT with cellular connectivity:
- Suppose the auto-connect is enabled on a cellular service.
- Verify that a cellular service isn't automatically connected before
  a user logs in, but can be connected manually.
- Verify that the cellular service is automatically connected after
  the user logs in.
- Disable auto-connect on the cellular service. Reboot the device.
  Verify that the cellular service isn't automatically connected before
  and after the user logs in, but can be connected manually.

Change-Id: I1da9ca577b0b3986b6fff666b5f9d25e187c300a
Reviewed-on: https://chromium-review.googlesource.com/851267
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>