midis: Fix upstart script
The midis daemon should be killed on ARC stopping, and should respawn
when the ARC instance starts up.

This patch fixes the upstart script so that the midis daemon behaves as
expected. This will also ensure that the daemon does not remain running
for accounts that don't have ARC enabled.

BUG=chromium:799170
TEST=midis daemon respawned successfully after sign-out / sign-in.

Change-Id: I042b84c62d2dbb215709fa8aad6107bd47d0daea
Reviewed-on: https://chromium-review.googlesource.com/853086
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>