arc-setup: Do not restore /var/run/arc/shared_mount's context
The directory is not a mount point on the host side. It's mkdir'ed in
arc-setup's OnSetup(), and is given the host-side parent directory's
SELinux label which is 'u:object_r:tmpfs:s0'. The default label is
good enough for the directory, and we don't need to run restorecon
in the OnPreChroot() hook.

If we do that, since the Android's file_contexts file doesn't have
an entry for the directory, the container's /var/run/arc's label,
which is 'u:object_r:rootfs:s0', is used for shared_mounts/, but the
label is not what we want to use for the directory.

BUG=b:71709605
BUG=b:69266495
TEST=cheets_SELinuxTest passes with nyc+run_oci
TEST=caroline-arcnext still works

Change-Id: I8feca88d68b18978673fc1043cd052e19d548809
Reviewed-on: https://chromium-review.googlesource.com/854685
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>