modemfwd: use equipment ID property rather than IMEI/MEID
If the SIM card is removed from a modem, then the IMEI/MEID fields
might not be set in the ModemManager object, and thus won't appear
in the expected places in shill either. However, the equipment
ID will still be populated, so that's a better place to find this
stable identifier.

CQ-DEPEND=CL:855534
BUG=None
TEST=run modemfwd with SIM card removed and check that equipment
  ID string logged to the console is not empty, then insert the SIM
  and verify that the equipment ID doesn't change

Change-Id: I908e6e2d547e66dfbe9da3d970b6181f092960b8
Reviewed-on: https://chromium-review.googlesource.com/855596
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>