metrics: limit daemon size and make it oom-killable
We set the maximum VM size of the metrics_daemon to 150MB.
Typical VM size on x86 is about 30MB.

We also make it oom-killable by setting its score adjustment to 0.
Normally Chrome OS sets this to -1000 (unkillable) for all system
processes.

This is motivated by a yet-to-be-rooted memory leak, but we feel
that it's not a bad idea in any case since no other processes
depend on the metrics daemon, and it doesn't have much state
so restarting it has small impact.

BUG=b:70771265
TEST=ran on device, checked /proc/<pid>/{limits,oom_score_adj}

Change-Id: Iaffd7fe0c028a2ef9ef90d4ed141296b21718a1e
Reviewed-on: https://chromium-review.googlesource.com/865538
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>