crash: distinguish service failures with exit status
Previously, all failures of a particular service are grouped
into the same "stable signature" of crash reports, which
is generated from the name of the service in question.

However, it can be useful to distinguish different reasons
for a service failure.  Thus, we make use of the service's
exit status and incorporate it into the stable signature.
The following format is used:

  stable_signature = {hash}-exit{exit_status}-{service_name}

BUG=b:69892693
TEST=cros_workon_make --board=poppy crash-reporter --test
TEST=USE=-cros-debug emerge-poppy -v --nodeps crash-reporter
     cros deploy --force --board=poppy dut crash-reporter
     stop anomaly-collector
     start anomaly-collector
     start hammerd EC_IMAGE_PATH=/
     cat /run/anomaly-collector/service-fail
       --> 2fdf045e-exit1-hammerd

Change-Id: I1c29bb621915a11b1a36d58f6cdb6d1c6e35033a
Reviewed-on: https://chromium-review.googlesource.com/866608
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>