chromeos-config: Drop the 'all models' option
This feature is not actually used, since we now have commands which return
the required configuration in packaged form for ebuilds. It doesn't make a
lot of sense to get the value for a particular config for all models. It
is a fairly low-level feature, requiring ebuilds to cope with duplicate
data when models have the same value.

Drop it.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge chromeos-config-host

Change-Id: I3959ce61b00e772fd22189e83e4c9da94255b2d1
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/892199
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>