init: Included TPM transition case when deciding to wipe
When the devmode is blocked by firmware management parameters
from TPM NVRAM, after switching from devmode to verified boot or
back the ${DEV_MODE_FILE} is not created.
To clear the stateful partitions in this case additional check
is added to see if the TPM is not owned but install_attributes
file is present. This means the device was owned and the TPM was
cleared.

BUG=chromium:701292
TEST=None

Change-Id: I6f2c83daf389d900faca8977f1386a94b19d1c63
Reviewed-on: https://chromium-review.googlesource.com/870312
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>