authpolicy: Gracefully handle missing LDAP attributes
Gracefully handle missing pwdLastSet and userAccountControl fields in
LDAP user queries. If the fields are missing, authpolicy won't be able
to detect expired passwords and assumes the password is valid. The user
will still be able to log into Chrome until the TGT expires. Then the
user is prompted to do an online login, which will trigger the password
change UI.

To limit the time until an online login is enforced, the user TGT is not
automatically renewed if the fields are missing. This means, the user
is prompted to relog after a day instead of after a week. This is
important for security reasons in case the admin wants the user to
change their password. We could also always force an online login, but
that would be extremely disruptive for the user (e.g. can never work
offline).

From the bug, it is unclear how the fields got missing.

BUG=chromium:795758
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy

Change-Id: I8331cd2137a583edcfa53b323acb60b388b57c8e
Reviewed-on: https://chromium-review.googlesource.com/870874
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>