cryptohome: remove captures from lambdas passed to base::Bind
Bind arguments by explicitly specifying them in the argument
list and using base::Bind's argument-binding functionality instead
of closing over them so that future versions of libchrome that
only allow captureless lambdas in Bind don't break.

One of the lambdas in question was extracted because it was
several lines long and closed over |this|, so it should just be
a private method.

BUG=b:37434548
TEST=unit tests

Change-Id: I709605033b7798019683c7063717d7df8ad8a6d7
Reviewed-on: https://chromium-review.googlesource.com/875247
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>