trunks: extract lambdas capturing the enclosing object
These lambdas are passed to base::Bind and are somewhat complicated
so it is better to extract them as private helper methods and
exchange the capture of |this| for binding |this| explicitly in
base::Bind. This also prepares us for the future libchrome uprev
because base::Bind will no longer accept lambdas that don't have
empty capture lists.

BUG=b:37434548
TEST=unit tests

Change-Id: Ia33da5601e8797b5dc098854824d83fa5bf78718
Reviewed-on: https://chromium-review.googlesource.com/875319
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>