vm_tools: concierge: Set pgid before starting crosvm
When crosvm detects that one or more of its child processes have not
exited in time it "fixes" the glitch by sending SIGKILL to every process
in its process group.  This ends up also killing vm_concierge.

Put crosvm in its own process group so that it cannot kill vm_concierge
in this way.

BUG=chromium:786663
TEST=stop a vm through concierge_client and see that vm_concierge does
     not get killed

Change-Id: Id2e78484b3b08dd3f09776730f01f70072fe3d3f
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/882192
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>