modemfwd: load text format protos
While we lose the ability to make all of the messages LITE, we
gain debuggability and easier test writing, as well as eliding
the proto compilation when packaging up firmware directories.
It also simplifies reading protos out of files since those no
longer require us to roll our own file reader stream.

This also changes the name of the manifest prototxt files so
that it's theoretically possible to use the same directory as
both the helper and firmware dir, which might be useful if you
want to send around one package during bringup, and using the
.prototxt extension makes it more obvious that this is a text
format proto file.

BUG=b:37953355
TEST=unit tests, deploy and load helper dir

Change-Id: I4705b5aa35d1dd8b0074fc36bed2808a1d06bf50
Reviewed-on: https://chromium-review.googlesource.com/885083
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>